semregbackend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
distlib==0.3.5
dj-database-url==0.5.0
Django==3.2.13
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-dotenv==1.4.2
django-extensions==3.2.0
django-filter==22.1
django-heroku==0.3.1
django-model-utils==4.2.0
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
importlib-metadata==4.12.0
package-json==0.0.0
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2021.5
virtualenv==20.15.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
wrapt==1.11.2
zipp==3.8.1
